# docs/swagger.yaml
basePath: /api
definitions:
  handlers.ActualizarProducto:
    properties:
      category:
        example: Electrónicos
        type: string
      description:
        example: Laptop HP con procesador Intel i7
        type: string
      name:
        example: Laptop HP Actualizada
        type: string
      price:
        example: 1299.99
        type: number
      sku:
        example: LAP-002
        type: string
    type: object
...
  handlers.CrearProducto:
    properties:
      category:
        example: Electrónicos
        type: string
      description:
        example: Laptop HP con procesador Intel i5
        type: string
      name:
        example: Laptop HP
        type: string
      price:
        example: 999.99
        type: number
      sku:
        example: LAP-001
        type: string
    required:
    - name
    - price
    - sku
    type: object
...
  handlers.Producto:
    description: Modelo de producto
    properties:
      category:
        example: Electrónicos
        type: string
      description:
        example: Laptop HP con procesador Intel i5
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Laptop HP
        type: string
      price:
        example: 999.99
        type: number
      sku:
        example: LAP-001
        type: string
    required:
    - name
    - price
    - sku
    type: object
  handlers.ProductoDetalle:
    properties:
      activo:
        example: true
        type: boolean
      category:
        example: Electrónicos
        type: string
      created_at:
        type: string
      description:
        example: Laptop HP con procesador Intel i5
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Laptop HP
        type: string
      price:
        example: 999.99
        type: number
      sku:
        example: LAP-001
        type: string
      updated_at:
        type: string
    required:
    - name
    - price
    - sku
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API para gestión de productos
  title: API de Productos
  version: "1.0"
paths:
  /ActivarDesactivarProductos:
    patch:
      consumes:
      - application/json
      description: Cambia el estado activo/inactivo de un producto
      parameters:
      - description: ID del producto
        in: query
        name: id
        required: true
        type: string
      - description: true para activar, false para desactivar
        in: query
        name: activate
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProductoDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activar/Desactivar producto
      tags:
      - productos
...
  /ActualizarProductos:
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un producto existente
      parameters:
      - description: ID del producto
        in: query
        name: id
        required: true
        type: string
      - description: Datos del producto
        in: body
        name: producto
        required: true
        schema:
          $ref: '#/definitions/handlers.ActualizarProducto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProductoDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar producto
      tags:
      - productos
...
  /CrearProductos:
    post:
      consumes:
      - application/json
      description: Crea un nuevo producto en el sistema
      parameters:
      - description: Datos del producto
        in: body
        name: producto
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearProducto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ProductoDetalle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear producto
      tags:
      - productos
...
  /EliminarProductos:
    delete:
      consumes:
      - application/json
      description: Elimina un producto del sistema
      parameters:
      - description: ID del producto
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar producto
      tags:
      - productos
...
  /ListarProductos:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de todos los productos activos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Producto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar productos
      tags:
      - productos
...
  /ObtenerProductos:
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de un producto específico
      parameters:
      - description: ID del producto
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProductoDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener producto por ID
      tags:
      - productos
...
schemes:
- http
swagger: "2.0"